split apps (news, comments, users)
