--- social_network/api/models.py
@@ -1,1 +1,0 @@
-from django.db import models
