+++ social_network/comments/urls.py
@@ -1,0 +1,7 @@
+from django.urls import path
+from .views import CommentsList
+
+
+urlpatterns = [
+    path('', CommentsList.as_view())
+]
