+++ social_network/user/serializers.py
@@ -1,0 +1,9 @@
+from rest_framework import serializers
+from .models import UserAPI
+
+
+class UserAPISerializer(serializers.ModelSerializer):
+
+    class Meta:
+        fields = ('id', 'username', 'email', 'role')
+        model = UserAPI
