+++ social_network/user/views.py
@@ -1,3 +1,13 @@
-from django.shortcuts import render
+from rest_framework import generics
+from .serializers import UserAPISerializer
+from .models import UserAPI
 
-# Create your views here.
+
+class UserList(generics.ListAPIView):
+    queryset = UserAPI.objects.all()
+    serializer_class = UserAPISerializer
+
+
+class UserDetail(generics.RetrieveUpdateDestroyAPIView):
+    queryset = UserAPI.objects.all()
+    serializer_class = UserAPISerializer
